Revert PriorityMergeReader to implement IPointReader
